+++ components/frontend/src/subject/SubjectTableFooter.test.js
@@ -6,7 +6,7 @@ import * as fetch_server_api from '../api/fetch_server_api';
 import { datamodel, report } from "../__fixtures__/fixtures";
 import { SubjectTableFooter } from "./SubjectTableFooter";
 
-const stopSorting = jest.fn()
+const stopFilteringAndSorting = jest.fn()
 
 jest.mock("../api/fetch_server_api.js")
 fetch_server_api.fetch_server_api = jest.fn().mockResolvedValue({ ok: true });
@@ -19,14 +19,14 @@ it('shows the add metric button and adds a metric when clicked', () => {
                     <SubjectTableFooter
                         subjectUuid="subject_uuid"
                         subject={report.subjects.subject_uuid}
-                        stopSorting={stopSorting} />
+                        stopFilteringAndSorting={stopFilteringAndSorting} />
                 </Table>
             </DataModel.Provider>
         </Permissions.Provider>
     );
     fireEvent.click(getByText(/Add metric/))
     fireEvent.click(screen.getByText(/Metric type/));
-    expect(stopSorting).toBeCalled()
+    expect(stopFilteringAndSorting).toBeCalled()
     expect(fetch_server_api.fetch_server_api).toHaveBeenCalledWith("post", "metric/new/subject_uuid", { type: "metric_type" });
 });
 
@@ -39,7 +39,7 @@ it('copies a metric when the copy button is clicked and a metric is selected', a
                         subjectUuid="subject_uuid"
                         subject={report.subjects.subject_uuid}
                         reports={[report]}
-                        stopSorting={stopSorting} />
+                        stopFilteringAndSorting={stopFilteringAndSorting} />
                 </Table>
             </DataModel.Provider>
         </Permissions.Provider>);
@@ -57,7 +57,7 @@ it('moves a metric when the move button is clicked and a metric is selected', as
                         subjectUuid="subject_uuid"
                         subject={report.subjects.subject_uuid}
                         reports={[report]}
-                        stopSorting={stopSorting} />
+                        stopFilteringAndSorting={stopFilteringAndSorting} />
                 </Table>
             </Permissions.Provider>
         </DataModel.Provider>)